+++ src/test/java/com/openggf/game/rewind/TestScalarOnlyCodecDeletion.java
@@ -526,6 +526,27 @@ public class TestScalarOnlyCodecDeletion {
                     "com.openggf.game.sonic2.objects.MonitorContentsObjectInstance",
                     GameId.S2));
 
+    private static final List<MutableFieldCoverageCandidate> BATCH33_MUTABLE_FIELDS =
+            List.of(
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic2.objects.MTZSpinTubeObjectInstance",
+                            "xFlipped"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic2.objects.MonitorContentsObjectInstance",
+                            "subtype"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic2.objects.SteamPuffObjectInstance",
+                            "xFlipped"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic2.objects.VerticalLaserObjectInstance",
+                            "posX", "posY"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic2.objects.badniks.BadnikProjectileInstance",
+                            "type"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic2.objects.badniks.SpikerDrillObjectInstance",
+                            "vFlip"));
+
     private static final List<CodecDeletionCandidate> BATCH31_DELETED_CODECS = List.of(
             new CodecDeletionCandidate(BombPrizeObjectInstance.class.getName(), GameId.S2));
 
@@ -3689,6 +3710,23 @@ public class TestScalarOnlyCodecDeletion {
     @Test
     void batch25ConstructorScalarsAreMutableForCompactRestore() {
         for (MutableFieldCoverageCandidate candidate : BATCH25_MUTABLE_FIELDS) {
+            Class<?> cls = loadClass(candidate.fqn());
+            for (String fieldName : candidate.fieldNames()) {
+                try {
+                    var field = findField(cls, fieldName);
+                    assertFalse(Modifier.isFinal(field.getModifiers()),
+                            cls.getName() + "#" + fieldName
+                                    + " must be mutable so compact restore can replay captured scalars");
+                } catch (NoSuchFieldException e) {
+                    throw new AssertionError("Missing scalar field " + cls.getName() + "#" + fieldName, e);
+                }
+            }
+        }
+    }
+
+    @Test
+    void batch33ConstructorScalarsAreMutableForCompactRestore() {
+        for (MutableFieldCoverageCandidate candidate : BATCH33_MUTABLE_FIELDS) {
             Class<?> cls = loadClass(candidate.fqn());
             for (String fieldName : candidate.fieldNames()) {
                 try {